backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
mysqlclient==2.0.3
numpy==1.19.5
opencv-python==4.5.1.48
pytz==2020.5
sqlparse==0.4.1
